+++ lib/com/vaadin/polymer/Polymer.java
@@ -6,7 +6,7 @@ import java.util.Set;
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.dom.client.Document;
 import com.google.gwt.dom.client.Element;
-import com.vaadin.polymer.webapi.Function;
+import com.vaadin.polymer.elemental.Function;
 
 public abstract class Polymer {
 
@@ -104,8 +104,8 @@ public abstract class Polymer {
     /**
      * Returns the JsInterop instance of Document
      */
-    public static com.vaadin.polymer.webapi.Document getDocument() {
-        return (com.vaadin.polymer.webapi.Document)Document.get();
+    public static com.vaadin.polymer.elemental.Document getDocument() {
+        return (com.vaadin.polymer.elemental.Document)Document.get();
     }
 
     /**
@@ -129,12 +129,12 @@ public abstract class Polymer {
         l.href = href;
         if (onload) {
           l.onload = function() {
-             onload.@com.vaadin.polymer.webapi.Function::call(*)();
+             onload.@com.vaadin.polymer.elemental.Function::call(*)();
           }
         }
         if (onerror) {
           l.onerror = function() {
-              onerror.@com.vaadin.polymer.webapi.Function::call(*)();
+              onerror.@com.vaadin.polymer.elemental.Function::call(*)();
           }
         }
         $doc.head.appendChild(l);
