+++ lib/com/vaadin/polymer/elemental/Document.java
@@ -1,4 +1,4 @@
-package com.vaadin.polymer.webapi;
+package com.vaadin.polymer.elemental;
 
 import com.google.gwt.core.client.js.JsProperty;
 import com.google.gwt.core.client.js.JsType;
