+++ lib/com/vaadin/polymer/elemental/Node.java
@@ -1,4 +1,4 @@
-package com.vaadin.polymer.webapi;
+package com.vaadin.polymer.elemental;
 
 import com.google.gwt.core.client.js.JsType;
